[c++,python] Remove tiledb-py import for all arrays
* Add vacuuming and consolidation to C++
* Remove TileDB ArraySchema creation
* Rename TileDBArray to SOMAArray to reflect that arrays are no
  longer TileDB-Py arrays but clib.SOMAArray
