+++ apis/python/src/tiledbsoma/_common_nd_array.py
@@ -7,31 +7,20 @@
 
 from typing import Optional, Sequence, Tuple, Union, cast
 
-import numpy as np
 import pyarrow as pa
 import somacore
 from somacore import options
 from typing_extensions import Self
 
-import tiledb
-
-from . import _arrow_types, _util
-from ._exception import (
-    AlreadyExistsError,
-    NotCreateableError,
-    is_already_exists_error,
-    is_not_createable_error,
-)
-from ._tiledb_array import TileDBArray
+from ._soma_array import SOMAArray
 from ._types import OpenTimestamp
 from .options._soma_tiledb_context import (
     SOMATileDBContext,
-    _validate_soma_tiledb_context,
 )
 from .options._tiledb_create_options import TileDBCreateOptions
 
 
-class NDArray(TileDBArray, somacore.NDArray):
+class NDArray(SOMAArray, somacore.NDArray):
     """Abstract base for the common behaviors of both kinds of NDArray."""
 
     __slots__ = ()
@@ -93,26 +82,7 @@ class NDArray(TileDBArray, somacore.NDArray):
         Lifecycle:
             Experimental.
         """
-        context = _validate_soma_tiledb_context(context)
-        schema = cls._build_tiledb_schema(
-            type,
-            shape,
-            TileDBCreateOptions.from_platform_config(platform_config),
-            context,
-            is_sparse=cls.is_sparse,
-        )
-        try:
-            handle = cls._create_internal(uri, schema, context, tiledb_timestamp)
-            return cls(
-                handle,
-                _dont_call_this_use_create_or_open_instead="tiledbsoma-internal-code",
-            )
-        except tiledb.TileDBError as tdbe:
-            if is_already_exists_error(tdbe):
-                raise AlreadyExistsError(f"{uri!r} already exists")
-            if is_not_createable_error(tdbe):
-                raise NotCreateableError(f"{uri!r} cannot be created")
-            raise
+        raise NotImplementedError("must be implemented by child class.")
 
     @property
     def shape(self) -> Tuple[int, ...]:
@@ -123,7 +93,7 @@ class NDArray(TileDBArray, somacore.NDArray):
         Lifecycle:
             Experimental.
         """
-        return cast(Tuple[int, ...], tuple(self._soma_reader().shape))
+        return cast(Tuple[int, ...], tuple(self._handle.shape))
 
     def reshape(self, shape: Tuple[int, ...]) -> None:
         """Unsupported operation for this object type.
@@ -132,77 +102,6 @@ class NDArray(TileDBArray, somacore.NDArray):
             Experimental.
         """
         raise NotImplementedError("reshape operation not implemented.")
-
-    @classmethod
-    def _build_tiledb_schema(
-        cls,
-        type: pa.DataType,
-        shape: Sequence[Union[int, None]],
-        create_options: TileDBCreateOptions,
-        context: SOMATileDBContext,
-        *,
-        is_sparse: bool,
-    ) -> tiledb.ArraySchema:
-        _util.check_type("type", type, (pa.DataType,))
-
-        if not pa.types.is_primitive(type):
-            raise TypeError(
-                f"Unsupported type {type} --"
-                " SOMA NDArrays only support primtive Arrow types"
-            )
-
-        if not shape:
-            raise ValueError("SOMA NDArrays must have a nonzero number of dimensions")
-
-        dims = []
-        for dim_idx, dim_shape in enumerate(shape):
-            dim_name = f"soma_dim_{dim_idx}"
-            dim_capacity, dim_extent = cls._dim_capacity_and_extent(
-                dim_name, dim_shape, create_options
-            )
-            dim = tiledb.Dim(
-                name=dim_name,
-                domain=(0, dim_capacity - 1),
-                tile=dim_extent,
-                dtype=np.int64,
-                filters=create_options.dim_filters_tiledb(
-                    dim_name,
-                    [
-                        dict(
-                            _type="ZstdFilter",
-                            level=create_options.sparse_nd_array_dim_zstd_level,
-                        )
-                    ],
-                ),
-            )
-            dims.append(dim)
-        dom = tiledb.Domain(dims, ctx=context.tiledb_ctx)
-
-        attrs = [
-            tiledb.Attr(
-                name="soma_data",
-                dtype=_arrow_types.tiledb_type_from_arrow_type(type),
-                filters=create_options.attr_filters_tiledb("soma_data", ["ZstdFilter"]),
-                ctx=context.tiledb_ctx,
-            )
-        ]
-
-        cell_order, tile_order = create_options.cell_tile_orders()
-
-        # TODO: accept more TileDB array-schema options from create_options
-        # https://github.com/single-cell-data/TileDB-SOMA/issues/876
-        return tiledb.ArraySchema(
-            domain=dom,
-            attrs=attrs,
-            sparse=is_sparse,
-            allows_duplicates=create_options.allows_duplicates,
-            offsets_filters=create_options.offsets_filters_tiledb(),
-            validity_filters=create_options.validity_filters_tiledb(),
-            capacity=create_options.capacity,
-            tile_order=tile_order,
-            cell_order=cell_order,
-            ctx=context.tiledb_ctx,
-        )
 
     @classmethod
     def _dim_capacity_and_extent(
